server/requirements.txt
django-rest-framework
pytest
ipdb
